requirements.txt
Django==1.8.3
django-tinymce==2.3.0
simplejson==3.8.2
pillow==3.2.0
Fabric==1.8.2
Jinja2==2.8
